Question parser added
Added a question parser, a questions xml, and the 'mainclass' file which
will be our main game class. all questions MUST be typed in the same
style that i used in the xml.
The xml reader doesn't recognize spaces so i used a minus sign instead
and removed/replaced it later otherwise this should work for us when we
get questions. the examples include a multi choice and a true or false
question.